Rename myapp to trips
