+++ apps/authentication/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
-from apps.myapp.models import Country
+from apps.trips.models import Country
 
 class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
