--- apps/myapp/templates/partials/cities.html
@@ -1,5 +1,0 @@
-<select name="city" class="w-full text-center font-serif bg-gray-50 border border-gray-300 text-black text-sm rounded-lg block p-2 mb-4">
-    {% for city in cities%}
-        <option value="{{city.pk}}">{{city.name}}</option>
-    {% endfor %}
-</select>
