+++ apps/trips/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class TripsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    label='trips'
+    name = 'apps.trips'
