+++ apps/trips/templates/configtrip.html
@@ -1,0 +1,95 @@
+{% extends "navbar.html" %}
+{% load static %}
+{% block mainbody %}
+
+<h1 class="text-xl font-normal font-serif text-black mb-4">
+    {{trip.title}}
+</h1>
+
+{% for destination in trip.destination_set.all %}
+    <div class="items-center p-4 bg-white rounded-lg shadow w-2/3 font-serif">
+        <div class="flex">
+            <div class="w-1/6">
+                <img src="https://flagcdn.com/h240/{{destination.country.alpha2code}}.png" class="my-2 w-12 h-8 object-cover">
+            </div>
+            <div class="w-2/6 flex flex-row items-center">
+                <img src="https://flagcdn.com/h240/{{destination.country.alpha2code}}.png" class="my-2 w-3 h-2 object-cover">
+                <p class="text-lg font-medium ml-2">{{ destination.city.name }}</p>
+            </div>
+            <div class="w-2/6 pl-4">
+                <div class="flex flex-row items-center">
+                    <svg xmlns="http://www.w3.org/2000/svg" width="10" height="10" fill="#5F9595" class="bi bi-calendar-event" viewBox="0 0 16 16">
+                        <path d="M11 6.5a.5.5 0 0 1 .5-.5h1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-1a.5.5 0 0 1-.5-.5v-1z"/>
+                        <path d="M3.5 0a.5.5 0 0 1 .5.5V1h8V.5a.5.5 0 0 1 1 0V1h1a2 2 0 0 1 2 2v11a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V3a2 2 0 0 1 2-2h1V.5a.5.5 0 0 1 .5-.5zM1 4v10a1 1 0 0 0 1 1h12a1 1 0 0 0 1-1V4H1z"/>
+                    </svg>
+                    <p class="text-sm text-gray-600 ml-2">{{ destination.start_date|date:"d M Y" }}</p>
+                </div>
+                <div class="flex flex-row items-center trans">
+                    <svg xmlns="http://www.w3.org/2000/svg" width="10" height="10" fill="#5F9595" class="bi bi-calendar-week" viewBox="0 0 16 16">
+                        <path d="M11 6.5a.5.5 0 0 1 .5-.5h1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-1a.5.5 0 0 1-.5-.5v-1zm-3 0a.5.5 0 0 1 .5-.5h1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-1a.5.5 0 0 1-.5-.5v-1zm-5 3a.5.5 0 0 1 .5-.5h1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-1a.5.5 0 0 1-.5-.5v-1zm3 0a.5.5 0 0 1 .5-.5h1a.5.5 0 0 1 .5.5v1a.5.5 0 0 1-.5.5h-1a.5.5 0 0 1-.5-.5v-1z"/>
+                        <path d="M3.5 0a.5.5 0 0 1 .5.5V1h8V.5a.5.5 0 0 1 1 0V1h1a2 2 0 0 1 2 2v11a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V3a2 2 0 0 1 2-2h1V.5a.5.5 0 0 1 .5-.5zM1 4v10a1 1 0 0 0 1 1h12a1 1 0 0 0 1-1V4H1z"/>
+                    </svg>
+                    <p class="text-sm text-gray-600 ml-2">{{ destination.end_date|date:"d M Y" }}</p>
+                </div>
+            </div>
+            <div class="w-1/6 flex justify-end items-center">
+                <form class="text-center mt-6 w-1/5" hx-post="{{trip.id}}" hx-target="body">
+                    <!-- Invisible field to store trip id and identify which form -->
+                    <input type="hidden" name="delete_destination_form" value="{{destination.id}}">
+                    <button>
+                        <svg hx-post="{{trip.id}}" class="hover:cursor-pointer" xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="#5F9595" class="bi bi-trash3-fill" viewBox="0 0 16 16">
+                            <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z"/>
+                        </svg>
+                    </button>
+                </form>
+            </div>
+        </div>
+        <div class="flex items-center justify-center">
+            <button class="text-xs font-medium text-custom-50 hover:underline hover:text-custom-100" _="on click toggle the *display of the next <div/> on click toggle the *display of the next <button/> on click toggle the *display of me">
+                See More
+            </button>
+            <button style="display: none;" class="text-xs font-medium text-custom-50 hover:underline hover:text-custom-100" _="on click toggle the *display of the next <div/> on click toggle the *display of the previous <button/> on click toggle the *display of me">
+                See Less
+            </button>
+        </div>
+        <div id="hi" style="display: none;" class="flex items-center justify-center mt-4">
+            <p>Some more information about the trip...</p>
+        </div>
+    </div>
+    <!-- Vertical line to connect to next destination -->
+    {% if not forloop.last %}
+        <div class="h-7 border-l border-gray-400"></div>
+        <div class="my-2">
+            <p class="text-sm font-light text-gray-500 text-center font-serif">
+                Travel not configured yet. <a href="" class="font-medium text-custom-50 hover:underline hover:text-custom-100">Configure here.</a>
+            </p>
+            <p class="text-sm underline font-medium text-gray-500 text-center font-serif">
+                Recommendation:
+            </p>
+        </div>
+        <div class="h-7 border-l border-gray-400"></div>
+    {% endif %}
+{% endfor %}
+
+<button _="on click toggle the *display of the next <form/> on click remove me" class="text-center mt-4 w-1/5 px-5 py-2 font-serif text-white bg-custom-50 hover:bg-custom-100 font-medium rounded-lg text-sm">Add a destination</button>
+
+<form style="display: none;" class="text-center mt-6 w-1/5" hx-post="{{trip.id}}" hx-target="body">
+    <!-- Invisible field identify which form -->
+    <input type="hidden" name="add_destination_form" value="1">
+    <select name="country" hx-get="{% url 'cities' %}" hx-trigger="change" hx-target="#cities" class="w-full text-center font-serif bg-gray-50 border border-gray-300 text-black text-sm rounded-lg block p-2 mb-4">
+        <option disabled selected>--- Select a country ---</option>
+        {% for country in countries%}
+            <option value="{{country.pk}}">{{country.name}}</option>
+        {% endfor %}
+    </select>
+    <div id="cities">
+        <select disabled name="city" class="w-full text-center font-serif bg-gray-50 border border-gray-300 text-black text-sm rounded-lg block p-2 mb-4">
+            <option>--- Select a city ---</option>
+        </select>
+    </div>
+    <input type="date" name="start_date" min="2023-01-01" placeholder="Start date" class="w-full text-center font-serif bg-gray-50 border border-gray-300 text-black text-sm rounded-lg block p-2 mb-4" />
+    <input type="date" name="end_date" min="2023-01-01" placeholder="End date" class="w-full text-center font-serif bg-gray-50 border border-gray-300 text-black text-sm rounded-lg block p-2 mb-4" />
+    <button class="px-5 py-2 font-serif w-full text-white bg-custom-50 hover:bg-custom-100 font-medium rounded-lg text-sm">Add another destination</button>
+</form>
+
+{% endblock %}
