+++ apps/trips/templates/mytrips.html
@@ -1,0 +1,49 @@
+{% extends "navbar.html" %}
+{% load static %}
+{% block mainbody %}
+
+<h1 class="text-xl font-normal font-serif text-black mb-4">
+    You have {{trips|length}} trip{% if trips|length > 1 %}s{% endif %} saved:
+</h1>
+
+{% for trip in trips %}
+<div class="flex flex-row w-2/3">
+    <div class="flex items-center p-4 bg-white rounded-lg shadow w-11/12 mb-4 font-serif hover:transform hover:cursor-pointer hover:scale-105 transition-all"
+    hx-get="trip/{{trip.id}}" hx-target="body" hx-push-url="true">
+        <div class="w-1/3">
+            <p class="text-lg font-medium">{{ trip.title }}</p>
+            <div class="flex">
+                {% for destination in trip.destination_set.all %}
+                    <img src="https://flagcdn.com/h240/{{ destination.country.alpha2code }}.png" class="w-6 h-4 mr-1">
+                {% endfor %}
+            </div>
+        </div>
+        <div class="w-2/3 pl-4">
+            <p class="text-sm text-gray-600">{{ trip.start_date|date:"d M Y" }}</p>
+            <p class="text-sm text-gray-600">{{ trip.end_date|date:"d M Y" }}</p>
+        </div>
+    </div>
+    <div class="flex items-center justify-center p-4 w-1/12 mb-4 font-serif">
+        <form hx-post="mytrips" hx-target="body">
+            <!-- Invisible field to store trip id and identify which form -->
+            <input type="hidden" name="delete_trip_form" value="{{trip.id}}">
+            <button>
+                <svg hx-post="{{trip.id}}" class="hover:cursor-pointer" xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="#5F9595" class="bi bi-trash3-fill" viewBox="0 0 16 16">
+                    <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z"/>
+                </svg>
+            </button>
+        </form>
+    </div>
+</div>
+{% endfor %}
+
+<button _="on click toggle the *display of the next <form/> on click remove me" class="text-center w-1/5 px-5 py-2 font-serif text-white bg-custom-50 hover:bg-custom-100 font-medium rounded-lg text-sm">Create a new trip</button>
+
+<form style="display: none;" class="text-center w-1/5" hx-post="" hx-target="body">
+    <!-- Invisible field identify which form -->
+    <input type="hidden" name="add_trip_form" value="1">
+    <input type="text" name="title" class="text-center font-serif bg-gray-50 border border-gray-300 text-black text-sm rounded-lg block w-full p-2 mb-4" placeholder="Trip Name">
+    <button class="px-5 py-2 font-serif w-full text-white bg-custom-50 hover:bg-custom-100 font-medium rounded-lg text-sm">Create trip</button>
+</form>
+
+{% endblock %}
